Rearrange code and add comprehensive comments
