+++ packages/exec/src/command-runner/middlware/pass-through.ts
@@ -1,0 +1,7 @@
+import {CommandRunnerMiddleware} from '../types'
+import {PromisifiedFn} from '../utils'
+
+/** Calls next middleware */
+export const passThrough: () => PromisifiedFn<CommandRunnerMiddleware> =
+  () => async (_, next) =>
+    next()
